tests/requirements.txt
aiohttp-devtools==0.13.1
aiounittest==1.4.0
bandit==1.6.2
behave==1.2.6
faker==4.1.1
lock-requirements==0.1.1
mypy==0.782
ptvsd==5.0.0a12
pydevd-pycharm==202.6397.98
pylint==2.5.3
requests==2.24.0
sort-requirements==1.3.0
unittest-data-provider==1.0.1
